Home/Index Page Part 1
Created the Nav bar with a custom made logo next to the website/company name.
Created also the carousel of images that represent the services that are being offered.
Started but not finished in creating fake reviews from made-up clients, trying to create the code of making the cards being next to each other and being of a reasonable size and being able to change size with the size of the webpage.