About Page in Progress + Contact Us Page Finished
Finished creating the Nav Bar on all pages with the connection between them so the user can change between one page and another. Also finished finding images that are going into the About Page the only thing left is to finish the code regarding the team itself then just add the footer that is already done form the Home Page.
The Contact Us Page is done as just added the already done nav bar and the pre-made Contact Us Form that we did in class, I did not change it as it was already perfect for what it was intended. Next I need to finish the About Page and Start and Finish the Services and Products Page.