Products In-Process
Manage to create a carousel that makes the art pop for the Art & Illustration, created sliding images that represent the posters, I made them different from the carousel so that the Products Webpage can pop.

I also managed to create a table to show how much the products cost using JavaScript. I will be using this same method for the Services Page.