Started and Finished the Service Page
Created a carousel in  the beginning to show what the "company" can offer. After I created a callery with the services offered and when the user clicks on it I made it transport you to the Service Page again it can be used to transport you its own page but not done.